requirement.txt
fastapi[standard]
sqlmodel
passlib[bcrypt]
pydantic-settings
bcrypt>=4.0,<5.0